pyproject.toml
[tool.poetry]
name = "forum"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.1"
flask-sqlalchemy = "^3.1.1"
psycopg2 = "^2.9.9"
python-slugify = "^8.0.1"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
